refactor: 테스트 하기 쉽도록 Repository 추상화, 데이터 생성자 주입
